Implement dynamic profile routes + refactor
