+++ web/src/components/ReviewTile.tsx
@@ -1,0 +1,12 @@
+export default () => {
+    return (
+        <div className="rounded-xl h-32 w-64 px-4 py-2 border border-secondary/20 bg-secondary/5 flex flex-col">
+            <h3 className="font-bold italic text-text/90">Pain 1993 - Drake</h3>
+            <p className="font-normal text-secondary/90 text-sm">Why lorem text placeholder...</p>
+            <div className="mt-auto flex flex-row place-content-between">
+                <p className="font-semibold text-text/80 text-md italic">24 Stars</p>
+                <p className="font-bold text-text/80 text-md italic ml-auto">125 Views</p>
+            </div>
+        </div>
+    );
+};
