+++ web/src/components/SidebarEntry.tsx
@@ -1,16 +1,11 @@
-import React from 'react';
-
-interface Song {
-  id: string;
-  title: string;
-  artist: string;
-}
+import React from "react";
+import {Song} from "@/types";
 
 interface SidebarEntryProps {
-  song: Song;
+    song: Song;
 }
 
-const SidebarEntry: React.FC<SidebarEntryProps> = ({ song }) => {
+const SidebarEntry: React.FC<SidebarEntryProps> = ({song}) => {
     return (
         <div className="flex flex-row place-content-before hover:bg-accent-neutral/20 transition px-2.5 py-1.5 my-0.5 rounded-md">
             <div className="w-5 h-5 bg-primary/20 my-auto"></div>
@@ -19,6 +14,6 @@ const SidebarEntry: React.FC<SidebarEntryProps> = ({ song }) => {
             </h3>
         </div>
     );
-}
+};
 
 export default SidebarEntry;
